requirements.txt
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-whois==0.8.0
shodan==1.29.1
colorama==0.4.6
rich==13.7.0
selenium==4.15.0
lxml==4.9.3